Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
litellm
python-dotenv
langgraph
langsmith
